+++ Sources/VComponents/Components/Modals/Modal/VModalUIModel.swift
@@ -55,7 +55,7 @@ public struct VModalUIModel {
 
     // MARK: Properties - Background
     /// Background color.
-    public var backgroundColor: Color = ColorBook.layer
+    public var backgroundColor: Color = ColorBook.background
 
     var groupBoxSubUIModel: VGroupBoxUIModel {
         var uiModel: VGroupBoxUIModel = .init()