+++ Sources/VComponents/Components/Modals/SideBar/VSideBarUIModel.swift
@@ -61,7 +61,7 @@ public struct VSideBarUIModel {
 
     // MARK: Properties - Background
     /// Background color.
-    public var backgroundColor: Color = ColorBook.layer
+    public var backgroundColor: Color = ColorBook.background
 
     var groupBoxSubUIModel: VGroupBoxUIModel {
         var uiModel: VGroupBoxUIModel = .init()